HW3/requirements.txt
Pillow==11.3.0
tqdm==4.67.1
tensorboard==2.20.0
torch>=2.3.0
torchvision>=0.18.0
torchaudio>=2.3.0